requirements.txt
Flask==1.0.3
Pillow==6.1.0
